requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
mssql-django==1.3
pyodbc==5.0.1
Pillow==10.1.0
django-admin-logs==1.0.0
django-otp==1.3.0 